# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
paths:
...
  '/api/properties/{property_id}':
    get:
      summary: ''
      operationId: getApiPropertiesProperty_id
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    agentId: 1
                    title: 'Maxime sint veritatis blanditiis excepturi.'
                    description: 'Eum velit sapiente sit eaque nam. Expedita vel molestias qui voluptatem. Sint fuga dolorem voluptates dolorum nam.'
                    price: '41133472.56'
                    propertyType: land
                    listingType: rent
                    bedrooms: 4
                    bathrooms: 2
                    size: 1202
                    address: '76611 Martin Trafficway'
                    area: nam
                    city: 'New Malika'
                    state: 'South Carolina'
                    status: sold
                    media: []
                properties:
                  data:
                    type: object
                    properties:
                      agentId:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: 'Maxime sint veritatis blanditiis excepturi.'
                      description:
                        type: string
                        example: 'Eum velit sapiente sit eaque nam. Expedita vel molestias qui voluptatem. Sint fuga dolorem voluptates dolorum nam.'
                      price:
                        type: string
                        example: '41133472.56'
                      propertyType:
                        type: string
                        example: land
                      listingType:
                        type: string
                        example: rent
                      bedrooms:
                        type: integer
                        example: 4
                      bathrooms:
                        type: integer
                        example: 2
                      size:
                        type: integer
                        example: 1202
                      address:
                        type: string
                        example: '76611 Martin Trafficway'
                      area:
                        type: string
                        example: nam
                      city:
                        type: string
                        example: 'New Malika'
                      state:
                        type: string
                        example: 'South Carolina'
                      status:
                        type: string
                        example: sold
                      media:
                        type: array
                        example: []
      tags:
        - 'Property management'
      security: []
    put:
      summary: ''
      operationId: putApiPropertiesProperty_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Property management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: integer
                  description: 'The <code>id</code> of an existing record in the agents table.'
                  example: 16
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                description:
                  type: string
                  description: ''
                  example: 'Eius et animi quos velit et.'
                price:
                  type: number
                  description: ''
                  example: 4326.41688
                property_type:
                  type: string
                  description: ''
                  example: apartment
                  enum:
                    - house
                    - apartment
                    - shortlet
                    - penthouse
                    - land
                    - commercial
                listing_type:
                  type: string
                  description: ''
                  example: rent
                  enum:
                    - sale
                    - rent
                status:
                  type: string
                  description: ''
                  example: unavailable
                  enum:
                    - available
                    - sold
                    - unavailable
                  nullable: true
                bedrooms:
                  type: integer
                  description: ''
                  example: 16
                  nullable: true
                bathrooms:
                  type: integer
                  description: ''
                  example: 16
                  nullable: true
                size:
                  type: number
                  description: ''
                  example: 4326.41688
                  nullable: true
                address:
                  type: string
                  description: 'Must not be greater than 500 characters.'
                  example: m
                area:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: i
                city:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: 'y'
                state:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: v
              required:
                - agent_id
                - title
                - description
                - price
                - property_type
                - listing_type
                - address
                - area
                - city
                - state
      security: []
    patch:
      summary: ''
      operationId: patchApiPropertiesProperty_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Property management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: integer
                  description: 'The <code>id</code> of an existing record in the agents table.'
                  example: 16
                title:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                description:
                  type: string
                  description: ''
                  example: 'Eius et animi quos velit et.'
                price:
                  type: number
                  description: ''
                  example: 4326.41688
                property_type:
                  type: string
                  description: ''
                  example: shortlet
                  enum:
                    - house
                    - apartment
                    - shortlet
                    - penthouse
                    - land
                    - commercial
                listing_type:
                  type: string
                  description: ''
                  example: rent
                  enum:
                    - sale
                    - rent
                status:
                  type: string
                  description: ''
                  example: available
                  enum:
                    - available
                    - sold
                    - unavailable
                  nullable: true
                bedrooms:
                  type: integer
                  description: ''
                  example: 16
                  nullable: true
                bathrooms:
                  type: integer
                  description: ''
                  example: 16
                  nullable: true
                size:
                  type: number
                  description: ''
                  example: 4326.41688
                  nullable: true
                address:
                  type: string
                  description: 'Must not be greater than 500 characters.'
                  example: m
                city:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: i
                state:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: 'y'
      security: []
    delete:
      summary: ''
      operationId: deleteApiPropertiesProperty_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Property management'
      security: []
    parameters:
      -
        in: path
        name: property_id
        description: 'The ID of the property.'
        example: 1
        required: true
        schema:
          type: integer
...